moved tensor api from ttnn to metal
